refactor(utils): extract context and validation into separate modules
- Move get_context() from utils/__init__.py to utils/context.py
- Move validate_email() from utils/__init__.py to utils/validation.py
- Re-export both functions from utils/__init__.py for backward compat
- Keeps the public API unchanged while improving module organization
