+++ chowkidar/utils/__init__.py
@@ -1,80 +1,5 @@
-from __future__ import annotations
-
-from re import match
-from typing import Any
-
-from graphql import GraphQLResolveInfo
-from django.http import HttpRequest
-from strawberry.types import Info
-from strawberry.django.context import StrawberryDjangoContext
-
-
-def get_context(info: HttpRequest | Info[Any, Any] | GraphQLResolveInfo) -> Any:
-    """Get the request context from the provided info object.
-
-    Parameters
-    ----------
-    info : HttpRequest | Info[Any, Any] | GraphQLResolveInfo
-        The info object from which to extract the request context.
-
-    Returns
-    -------
-    Any
-        The request context, which can be an HttpRequest or a StrawberryDjangoContext.
-
-    Raises
-    ------
-    TypeError
-        If the provided info object is not of a recognized type.
-
-    Notes
-    -----
-    - If the info object is an instance of HttpRequest, it is returned directly.
-    - If the info object is an instance of Info, its context attribute is checked.
-      - If the context is an instance of StrawberryDjangoContext, its request attribute is returned.
-      - Otherwise, the context itself is returned.
-    - If the info object is an instance of GraphQLResolveInfo, its context attribute is returned.
-    - If the info object is of an unrecognized type, a TypeError is raised.
-
-    """
-    if hasattr(info, "context"):
-        ctx = getattr(info, "context")
-        if isinstance(ctx, StrawberryDjangoContext):
-            return ctx.request
-
-        return ctx
-    return info
-
-
-def validate_email(email: str) -> str:
-    """Validate an email address against a basic RFC-like regex pattern.
-
-    Parameters
-    ----------
-    email : str
-        The email address string to validate.
-
-    Returns
-    -------
-    str
-        The validated email address, returned unchanged.
-
-    Raises
-    ------
-    AuthError
-        If the email address does not match the expected pattern (code ``INVALID_EMAIL``).
-
-    """
-    from chowkidar.utils.exceptions import AuthError
-
-    if not match(r"(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)", email):
-        raise AuthError(
-            message="You have entered an invalid email address",
-            code="INVALID_EMAIL",
-        )
-
-    return email
-
+from .context import get_context
+from .validation import validate_email
 
 __all__ = [
     "get_context",
